requirement.txt
boto3>=1.28.57
awscli>=1.29.57
botocore>=1.31.57
langchain==0.0.309
transformers>=4.24,<5
sqlalchemy
faiss-cpu>=1.7,<2
pypdf>=3.8,<4
pinecone-client
apache-beam
datasets
tiktoken
ipywidgets>=7,<8
matplotlib
xmltodict==0.13.0
duckduckgo-search
yfinance
pandas_datareader
langchain_experimental
pysqlite3
google-search-results
beautifulsoup4
pillow>=9.5,<10
pypdf>=3.8,<4
streamlit
